libraries updates and platformio fixes

+++ sensors_network/lib/MySensors/examples/DustSensor/DustSensor.ino
@@ -45,7 +45,6 @@
 #define MY_RADIO_NRF24
 //#define MY_RADIO_RFM69
 
-#include <SPI.h>
 #include <MySensors.h>  
 
 #define CHILD_ID_DUST 0
@@ -94,7 +93,7 @@ void loop() {
   Serial.println(dustDensity); // unit: ug/m3
  
   if (ceil(dustDensity) != lastDUST) {
-      send(dustMsg.set((int)ceil(dustDensity)));
+      send(dustMsg.set((int16_t)ceil(dustDensity)));
       lastDUST = ceil(dustDensity);
   }
